pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "oxidizedvision"
version = "1.0.1"
authors = [
  { name="Avaya Aggarwal", email="aggarwal.avaya27@gmail.com" },
]
description = "Compile your vision models to Rust for ultra-fast inference."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
]
keywords = ["machine-learning", "pytorch", "onnx", "rust", "inference", "deployment"]
dependencies = [
    "torch>=1.12",
    "torchvision",
    "onnx>=1.12",
    "onnx-simplifier",
    "onnxruntime>=1.12",
    "numpy>=1.21",
    "typer>=0.7",
    "pydantic>=1.10",
    "pyyaml>=6.0",
    "rich>=12.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "isort>=5.12",
    "ruff>=0.1",
    "mypy>=1.0",
]
quantization = [
    "onnxruntime-extensions",
]

[project.urls]
"Homepage" = "https://github.com/onepunchmonk/Oxidized-Vision"
"Bug Tracker" = "https://github.com/onepunchmonk/Oxidized-Vision/issues"
"Documentation" = "https://github.com/onepunchmonk/Oxidized-Vision/tree/main/docs"

[project.scripts]
oxidizedvision = "oxidizedvision.cli:main"

[tool.setuptools.packages.find]
where = ["python_client"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["python_client/tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
